Change `collection` behaviour.
It is changed to use the most performant way to associate many objects in a NoSQL database. 
Now an array of ids is persisted in order to retrieve all the objects without using an index search.
But it won't change the referenced object, the parent attribute must be set if is pretended to be persisted.

The previous behaviour was the relation-way to do it, using indices.